Changed locking system so that only one lock is ever in place on a given ticket and, RT::Ticket::Unlock only unlocks if the lock was an auto lock or Unlock was called with a true value (hard unlock) RT::Ticket::Lock now has a parameter to determine whether the lock is a hard (manual) lock or an auto lock
git-svn-id: svn+ssh://svn.bestpractical.com/svn/bps-public/rt/branches/3.7-EXPERIMENTAL-TUNIS@8358 e417ac7c-1bcc-0310-8ffa-8f5827389a85

+++ lib/RT/Ticket.pm
@@ -236,6 +236,7 @@ sub Locked {
 
 sub Lock {
     my $ticket = shift;
+    my $hard = shift;
 
     if ( $ticket->Locked() ) {
         return undef;
@@ -244,8 +245,8 @@ sub Lock {
             Name    => 'RT_Lock',
             Content => {
                 User      => $ticket->CurrentUser->id,
-                Timestamp => time()
-
+                Timestamp => time(),
+                Hard => $hard
             }
         );
     }
@@ -254,10 +255,12 @@ sub Lock {
 
 sub Unlock {
     my $ticket = shift;
+    my $hard = shift;
 
     my $lock = $ticket->RT::Ticket::Locked();
-     return undef unless $lock;
-     return undef unless $lock->Content->{User} ==  $ticket->CurrentUser->id;
+    return undef unless $lock;
+    return undef unless $lock->Content->{User} ==  $ticket->CurrentUser->id;
+    return undef if $lock->Content->{'Hard'} && !$hard;
     $ticket->DeleteAttribute('RT_Lock');
 }
